starknet: move 'types' to 'core' directory
