+++ packages/starknet/lib/src/contract/erc20.dart
@@ -1,6 +1,6 @@
 import 'package:starknet_provider/starknet_provider.dart';
 
-import '../types/index.dart';
+import '../core/types/index.dart';
 import 'contract.dart';
 
 class ERC20 extends Contract {
