+++ packages/starknet/lib/src/contract/model/compiled_contract.dart
@@ -5,7 +5,7 @@ import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:starknet_provider/starknet_provider.dart';
 
 import '../../crypto/index.dart';
-import '../../types/index.dart';
+import '../../core/types/index.dart';
 import './compiled_class_hash.dart';
 import 'contract_abi.dart';
 import 'entry_points_by_type.dart';
