+++ packages/starknet/lib/src/crypto/keccak.dart
@@ -5,7 +5,7 @@ import 'dart:typed_data';
 import 'package:pointycastle/digests/keccak.dart';
 
 import '../convert.dart';
-import '../types/index.dart';
+import '../core/types/index.dart';
 
 /// [mask250] is used to make sure a big int has less than 250 digits.
 final mask250 = BigInt.two.pow(250) - BigInt.one;
