+++ packages/starknet/lib/src/signer/base_signer.dart
@@ -1,6 +1,6 @@
 // TODO: should be moved  to "core" package
 
-import '../types/felt.dart';
+import '../core/types/felt.dart';
 
 abstract class BaseSigner {
   Future<List<Felt>> sign(BigInt messageHash, BigInt? seed);
