+++ packages/starknet/lib/src/snip/snip12/typed_data/typed_data.dart
@@ -7,7 +7,7 @@ import 'dart:convert';
 import 'package:freezed_annotation/freezed_annotation.dart';
 
 import '../../../crypto/index.dart';
-import '../../../types/index.dart';
+import '../../../core/types/index.dart';
 import 'encode.dart';
 import 'num.dart';
 import 'shortstring.dart';
